Initial test engine refactor to clean up code / separation of concerns
