+++ README.md
@@ -49,7 +49,7 @@ Add the Junit 5 and Devopology test engine jar dependencies...
 <dependency>
     <groupId>org.devopology</groupId>
     <artifactId>test-engine</artifactId>
-    <version>0.0.5</version>
+    <version>0.0.6-SNAPSHOT</version>
 </dependency>
 ```
