+++ src/main/resources/devopology-test-engine.properties
@@ -1,0 +1,1 @@
+version=v0.0.6-SNAPSHOT
