--- src/main/resources/test-engine.properties
@@ -1,1 +1,0 @@
-version=v0.0.5
