+++ src/test/java/org/devopology/test/engine/test/example/CamelCaseFunctionTest.java
@@ -10,6 +10,7 @@ import java.util.Collection;
 import java.util.function.Function;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Fail.fail;
 
 /**
  * Example test engine test... only runs from an IDE or via the test engine ConsoleRunner
@@ -82,6 +83,12 @@ public class CamelCaseFunctionTest {
 
     @Test
     public void test() {
-        assertThat(FUNCTION.apply(parameter.input)).isEqualTo(parameter.expected);
+        String actual = FUNCTION.apply(parameter.input);
+        System.out.println("test() input [" + parameter.input + "] expected [" + parameter.expected + "] actual [" + actual + "]");
+        assertThat(actual).isEqualTo(parameter.expected);
+
+        if (parameter.input.equals("THIS") && parameter.expected.equals("this")) {
+        //    fail("Forced failure");
+        }
     }
 }
